Fix DocVideo issue
